requirements.txt
requests==2.9.1
pymongo==3.5.1
redis==2.10.5
beautifulsoup4==4.6.0
PyYAML==3.12
git+git://github.com/scrapinghub/python-simhash#egg=python-simhash
lxml==4.0.0
mock==2.0.0
mongomock==3.10.0
freezegun==0.3.10
mockredispy==2.9.3
green==2.12.1
